requirements.txt
torch~=2.1.1
numpy~=1.26.2
opencv-python~=4.8.1.78
Pillow==10.1.0
tqdm~=4.65.2
albumentations~=1.3.1